FruitTransaction POJO
